Cleanup cache
# Conflicts:
#	packages/api/internal/nomad/instance_cache.go
